diet_project/requirements.txt
channels==2.4.0
channels-redis==2.4.2

daphne==2.5.0
Django==3.0.5
django-braces==1.12.0
django-common-helpers==0.9.2
django-cron==0.5.1
django-import-export==2.3.0
django-mathfilters==1.0.0
django-model-utils==2.6.1
djangorestframework==3.11.0
djongo==1.3.3
Pillow
pycrypto
mysqlclient
